moving rfm tests into ran tests
