+++ test/gen_ground_truth.sh
@@ -5,6 +5,10 @@ rm data/*.json
 ../dwig.py -pp -cd  1 -tl -rs 0 -os ran -s 4 > data/ran4_i_0.json
 ../dwig.py -pp -cd 12 -tl -rs 0 -os ran > data/ran1_i_3.json
 
+../dwig.py -pp -cd 1 -tl -rs 0 ran -a 0.0 > data/rfm1_i_1.json
+../dwig.py -pp -cd 1 -tl -rs 0 ran -a 0.0 -f > data/rfm1_i_2.json
+../dwig.py -pp -cd 1 -tl -rs 0 ran -a 0.0 -f -s 4 > data/rfm4_i_1.json
+
 ../dwig.py -pp -cd  2 -tl -rs 0 fl -sgs > data/fl_i_1.json
 ../dwig.py -pp -cd  2 -tl -rs 0 fl -sgs -s 4 -a 0.3 > data/fl_i_2.json
 ../dwig.py -pp -cd  2 -tl -rs 0 fl -sgs -mc -mll 0 > data/fl_i_3.json
@@ -20,6 +24,3 @@ rm data/*.json
 ../dwig.py -pp -cd 9 -tl -rs 0 wscn > data/wscn_i_2.json
 ../dwig.py -pp -cd 12 -tl -rs 0 wscn > data/wscn_i_3.json
 
-../dwig.py -pp -cd 1 -tl -rs 0 ran -a 0.0 > data/rfm1_i_1.json
-../dwig.py -pp -cd 1 -tl -rs 0 ran -a 0.0 -f > data/rfm1_i_2.json
-../dwig.py -pp -cd 1 -tl -rs 0 ran -a 0.0 -f -s 4 > data/rfm4_i_1.json
